Filter compiler flags not understood by clang-tidy
In the Key4hep stack we build Geant4 with IPO. Geant4 injects these
flags into the build when doing so and when it has been built with
gcc (regardless of what is actually being done by the target that is
linke against it). Since we don't want to build anything here, we just
filter these flags from the compile_commands.json database in order to
make clang-tidy not encounter these flags.